In printfs, cast size_ts to int in printf to avoid 64-bit warnings
Monotone-Parent: cc94d6953f36a7c89a0dc1510689638ed4bf4aff
Monotone-Revision: 13af8679c7ed8903a6c615365ce5089405e82ecf

+++ test/gen_test_vec.cpp
@@ -8,7 +8,7 @@ using fecpp::byte;
 void print_block(size_t block_no, size_t n,
                  const byte share[], size_t share_len)
    {
-   printf("block_%d=", block_no);
+   printf("block_%d=", (int)block_no);
    for(size_t j = 0; j != share_len; ++j)
       printf("%02X", share[j]);
    if(block_no != n - 1)
